requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
black==19.10b0
bleach==3.1.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
ciscoconfparse==1.5.1
click==7.1.1
colorama==0.4.3
cryptography==2.9
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
junos-eznc==2.2.1
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
mypy-extensions==0.4.3
napalm==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
ncclient==0.6.7
netaddr==0.7.19
netmiko==2.4.2
nornir==2.4.0
notebook==6.0.3
nxapi-plumbing==0.5.2
packaging==20.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
passlib==1.7.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
pyeapi==0.8.3
Pygments==2.6.1
pyIOSXR==0.53
pylint==2.4.4
PyNaCl==1.3.0
pynetbox==4.3.0
pyparsing==2.4.6
pyrsistent==0.15.7
pyserial==3.4
pytest==5.4.1
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scp==0.13.2
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.1
terminado==0.8.3
testpath==0.4.4
textfsm==1.1.0
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
